+++ server/services/monuments.js
@@ -10,7 +10,7 @@ const { pool } = require('../db');
 let logGPActivity;
 try { logGPActivity = require('../db').logGPActivity; } catch (_) {}
 let notifyPlayer;
-try { notifyPlayer = require('./notifications').notifyPlayer; } catch (_) {}
+try { notifyPlayer = require('../db').notifyPlayer; } catch (_) {}
 let newsService;
 try { newsService = require('./news'); } catch (_) {}
 let divSvc;